refactor(core): modulariza UTMs, HTTP, preços/moeda e lógica do BetaSignup
Extrai useBetaSignupForm; tipa analytics (track<E>); ativa code-splitting e ajusta chamadas de track().
