+++ src/components/BetaSignupForm.tsx
@@ -1,99 +1,17 @@
-import { useEffect, useMemo, useRef, useState } from 'react';
+import { useMemo } from 'react';
 import { useLocation } from 'react-router-dom';
-import { trackEvent } from '../lib/analytics';
+import { formatBRL } from '../lib/utils';
+import { DEFAULT_FOUNDER_DISCOUNT, PRICE_BUSINESS, PRICE_STARTER } from '../lib/pricing';
+import { useBetaSignupForm } from '../hooks/useBetaSignupForm';
 
 /** Props for the signup form component */
 type Props = { variant: "A" | "B"; selectedPlan?: 'starter' | 'business' | '' };
 /** Possible plan hints passed via query string */
 type PlanHint = 'starter' | 'business' | 'founder' | '';
-/** Plans the user can select */
-type Plan = 'starter' | 'business' | '';
-
-/** Keys for UTM capture persisted at root resolver */
-const UTM_STORAGE_KEY = 'chavexls_utm';
-const UTM_KEYS = [
-  'utm_source',
-  'utm_medium',
-  'utm_campaign',
-  'utm_term',
-  'utm_content',
-  'gclid',
-  'fbclid',
-  'ts', // timestamp of capture
-] as const;
-type UTMKey = typeof UTM_KEYS[number];
-
-/** Safely read UTMs from localStorage */
-function readStoredUtms(): Partial<Record<UTMKey, string>> {
-  try {
-    const raw = localStorage.getItem(UTM_STORAGE_KEY);
-    if (!raw) return {};
-    const obj = JSON.parse(raw) as Record<string, string>;
-    const picked: Partial<Record<UTMKey, string>> = {};
-    for (const k of UTM_KEYS) {
-      if (obj[k]) picked[k] = String(obj[k]);
-    }
-    return picked;
-  } catch {
-    return {};
-  }
-}
-
-/** Best-effort GA client id extraction (fallback-friendly) */
-function getClientId(): { cid: string; source: 'gtag'|'cookie'|'fallback' } {
-  // 1) Try gtag client id if available
-  try {
-    // ts-expect-error gtag global may exist
-    if (typeof window.gtag === 'function') {
-      // We don't know the measurement id here; keep a simple fallback to cookie
-      // (If needed later, we can wire a proper async gtag('get', 'G-XXXX', 'client_id', cb))
-    }
-  } catch { /* ignore */ }
-
-  // 2) Parse cookies (_ga or _ga_*). Formats vary between GA3/GA4; we keep raw id if unsure.
-  try {
-    const cookies = document.cookie.split(';').map(s => s.trim());
-    // _ga=GA1.2.XXXXXXXXXX.YYYYYYYYYY
-    const ga = cookies.find(c => c.startsWith('_ga='));
-    if (ga) {
-      const val = decodeURIComponent(ga.split('=')[1] || '');
-      const parts = val.split('.');
-      // Try to combine the last two numeric parts as cid; otherwise keep raw
-      if (parts.length >= 4) {
-        const cid = `${parts[2]}.${parts[3]}`;
-        if (/^\d+\.\d+$/.test(cid)) return { cid, source: 'cookie' };
-      }
-      return { cid: val, source: 'cookie' };
-    }
-    // _ga_<MEASUREMENT>=GS1.1.x.y...
-    const ga4 = cookies.find(c => c.startsWith('_ga_'));
-    if (ga4) {
-      const val = decodeURIComponent(ga4.split('=')[1] || '');
-      return { cid: val, source: 'cookie' };
-    }
-  } catch { /* ignore */ }
-
-  // 3) Fallback: ephemeral pseudo-id
-  return { cid: `lf_${Date.now()}_${Math.floor(Math.random()*1e6)}`, source: 'fallback' };
-}
-
-/** Format a number as BRL currency */
-function formatBRL(n: number): string {
-  try {
-    return n.toLocaleString('pt-BR', {
-      style: 'currency',
-      currency: 'BRL',
-      minimumFractionDigits: 2,
-    });
-  } catch {
-    return `R$ ${n.toFixed(2)}`;
-  }
-}
+// plan type moved to hook
 
 /** Pricing constants */
-const PRICE_STARTER = 19.9;
-const PRICE_BUSINESS = 59.9;
-const FOUNDER_DISCOUNT = 0.3; // 30%
+const FOUNDER_DISCOUNT = DEFAULT_FOUNDER_DISCOUNT; // 30%
 
 /** Beta signup form component */
 export default function BetaSignupForm({ variant, selectedPlan = '' }: Props) {
@@ -104,27 +22,17 @@ export default function BetaSignupForm({ variant, selectedPlan = '' }: Props) {
     return hint === 'starter' || hint === 'business' || hint === 'founder' ? hint : '';
   }, [location.search]);
 
-  // Selected plan visible to the user. If plan_hint is starter or business, preselect it.
-  const [plan, setPlan] = useState<Plan>('');
-  useEffect(() => {
-    if (planHint === 'starter' || planHint === 'business') setPlan(planHint);
-  }, [planHint]);
-
-  // If selectedPlan prop changes (from URL param in VariantA), update plan state
-  useEffect(() => {
-    if (selectedPlan === 'starter' || selectedPlan === 'business') setPlan(selectedPlan);
-  }, [selectedPlan]);
-
-  const [submitted, setSubmitted] = useState(false);
-
-  // Track first real interaction (for anti-bot delay)
-  const [started, setStarted] = useState(false);
-  const startedAtRef = useRef<number | null>(null);
-
-  // Optional message
-  const [message, setMessage] = useState('');
-  const messageLen = message.length;
-  const hasPhone = /\d{8,13}/.test(message.replace(/\D/g, '')); // 8–13 digits
+  const {
+    plan,
+    submitted,
+    message,
+    setMessage,
+    utms,
+    cid,
+    handleFirstFocus,
+    handlePlanChange,
+    onSubmit,
+  } = useBetaSignupForm({ variant, planHint, selectedPlanProp: selectedPlan });
 
   // Founder pricing is shown only for Variant A
   const founderMode = variant === 'A';
@@ -135,113 +43,7 @@ export default function BetaSignupForm({ variant, selectedPlan = '' }: Props) {
     ? { main: PRICE_BUSINESS * (1 - FOUNDER_DISCOUNT), original: PRICE_BUSINESS }
     : { main: PRICE_BUSINESS, original: null as number | null };
 
-  // Read UTMs and CID once (kept for hidden inputs and payload)
-  const utms = useMemo(() => readStoredUtms(), []);
-  const { cid } = useMemo(() => getClientId(), []);
-
-  /** Encode object to x-www-form-urlencoded for Netlify */
-  function encode(data: Record<string, string>): string {
-    return Object.keys(data)
-      .map((k) => encodeURIComponent(k) + '=' + encodeURIComponent(data[k]))
-      .join('&');
-  }
-
-  /** Track form_start once on first user focus */
-  function handleFirstFocus(): void {
-    if (!started) {
-      setStarted(true);
-      startedAtRef.current = Date.now();
-      trackEvent('form_start', { variant });
-    }
-  }
-
-  /** Handle plan selection changes and track analytics */
-  function handlePlanChange(next: Plan): void {
-    if (next !== plan) {
-      setPlan(next);
-      if (next) trackEvent('plan-select-change', { variant, plan: next });
-    }
-  }
-
-  /** Submit the form and send analytics events */
-  async function onSubmit(e: React.FormEvent<HTMLFormElement>): Promise<void> {
-    e.preventDefault();
-
-    // Anti-bot: require 3s since first interaction
-    const now = Date.now();
-    if (startedAtRef.current && now - startedAtRef.current < 3000) {
-      alert('Por favor, revise e envie novamente em alguns segundos.');
-      return;
-    }
-
-    const form = e.currentTarget;
-    const formData = new FormData(form);
-
-    const profile = String(formData.get('profile') || '');
-    const volume = String(formData.get('monthlyVolume') || '');
-    const accepted = formData.get('acceptedBetaTerms') ? 'yes' : 'no';
-    const selectedPlanValue = String(formData.get('plan') || '');
-    const messageValue = String(formData.get('message') || '');
-
-    // Build Netlify payload
-    const payload: Record<string, string> = {
-      'form-name': 'beta-signup',
-      name: String(formData.get('name') || ''),
-      email: String(formData.get('email') || ''),
-      profile,
-      monthlyVolume: volume,
-      acceptedBetaTerms: accepted,
-      variant,
-      plan: selectedPlanValue,
-      plan_hint: planHint,
-      message: messageValue,
-      // Tracking extras
-      cid,
-    };
-
-    // Append UTMs if present
-    for (const k of UTM_KEYS) {
-      const v = (utms as any)[k];
-      if (v) payload[k] = String(v);
-    }
-
-    try {
-      await fetch('/', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-        body: encode(payload),
-      });
-
-      // GA: DO NOT send the message text, only derived metrics
-      trackEvent('lead_submit', {
-        variant,
-        profile,
-        volume,
-        plan_hint: planHint,
-        message_len: Math.min(messageLen, 500), // max 500 chars
-        has_phone: hasPhone,
-        cid,
-        ...utms,
-      });
-      // Optional alias event (keep if you want both marked as conversion)
-      trackEvent('beta_signup', {
-        variant,
-        profile,
-        volume,
-        plan_hint: planHint,
-        message_len: Math.min(messageLen, 500),
-        has_phone: hasPhone,
-        cid,
-        ...utms,
-      });
-
-      setSubmitted(true);
-      form.reset();
-      setMessage('');
-    } catch {
-      alert('Falha no envio. Por favor, tente novamente.');
-    }
-  }
+  // Logic moved to useBetaSignupForm
 
   if (submitted) {
     return (
@@ -272,7 +74,7 @@ export default function BetaSignupForm({ variant, selectedPlan = '' }: Props) {
       <input type="hidden" name="plan_hint" value={planHint} />
       <input type="hidden" name="cid" value={cid} />
       {/* UTM hidden fields (only rendered if present) */}
-      {UTM_KEYS.map((k) =>
+      {Object.keys(utms).map((k) =>
         (utms as any)[k] ? <input key={k} type="hidden" name={k} value={(utms as any)[k]} /> : null
       )}
       {/* Honeypot */}
