+++ src/lib/pricing.ts
@@ -1,0 +1,8 @@
+/** Centralized pricing constants to avoid drift */
+export const PRICE_STARTER = 19.9;
+export const PRICE_BUSINESS = 59.9;
+/** Default founder discount (30%) used in Variant A displays */
+export const DEFAULT_FOUNDER_DISCOUNT = 0.3;
+
+export type PlanId = 'starter' | 'business';
+
